Add an entrance playthrough to make ER seeds easier to figure out
This adds a part in the playthrough which shows all randomized entrances unlocked in each playthrough sphere. It's meant to be an extension of the usual playthrough to help with reading ER seeds, not a standalone entrance playthrough. 
For this reason, it's synced with the normal playthrough and there might be spheres where no entrances are unlocked.

However, it should probably be improved to only list entrances that are actually useful to the playthrough, but that requires a lot more work and generation time so I'm not sure if it's actually worth doing. 
I think this should already be useful source of information in this state so it's worth adding in anyway.